+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import App from "./components";
+import App from "./App";
 import { AppContext, context } from "./store/index";
 
 const { Provider } = AppContext
